+++ server/core/services/__init__.py
@@ -1,5 +1,5 @@
 from .rcrainfo_client import RcraClient, get_rcra_client
-from .task_service import TaskService, get_task_status, launch_example_task
+from .task_status import TaskService, get_task_status, launch_example_task
 
 __all__ = [
     "RcraClient",
